finally fixed dashboard pages both stats and chart update live at the same time in sync
