+++ Frontend/src/lib/Components/Chart.svelte
@@ -5,53 +5,59 @@
 
   let chart;
   let chartCanvas;
-  let data; 
+  export let data = {}; 
   let interval; 
 
   onMount(() => {
     
     if (typeof window !== 'undefined') { // Check if running in the browser
       chart = createChart(chartCanvas, sampleChartConfig);
-      fetchData(); 
-      interval = setInterval(fetchData, 1000);
+      // fetchData(); 
+      updateChart();
+      // interval = setInterval(updateChart, 1000);
     }
 
     return () => {
       if (chart) {
         //chart.destroy();
       }
-      clearInterval(interval);
+      // clearInterval(interval);
     };
   });
 
   
 
-  async function fetchData() {
+  // async function fetchData() {
 
-      try {
-        const response = await fetch("http://localhost:8000/overview", {
-      method: "POST", 
-      headers: {
-        'Content-Type': 'application/json' 
-      }
+  //     try {
+  //       const response = await fetch("http://localhost:8000/overview", {
+  //     method: "POST", 
+  //     headers: {
+  //       'Content-Type': 'application/json' 
+  //     }
 
-    });
-        console.log("request being sent...");
-        // const response = fetch("http://localhost:8000");
-        const data = await response.json();
-        console.log(data);
-        //Voltage 1,2,3 as well as price
-        updateChart(data.Phase1, data.Phase2);
-      } catch (error) {
-        console.log("There was an error fetching the JSON for the chart..", error);
-      }
-  };
+  //   });
+  //       console.log("request being sent...");
+  //       // const response = fetch("http://localhost:8000");
+  //       const data = await response.json();
+  //       console.log(data);
+  //       //Voltage 1,2,3 as well as price
+  //       updateChart(data.Phase1, data.Phase2);
+  //     } catch (error) {
+  //       console.log("There was an error fetching the JSON for the chart..", error);
+  //     }
+  // };
 
-  function updateChart(data1, data2){
+    $: if (chart && Object.keys(data).length) {
+    updateChart();
+  }
+
+  function updateChart(){
 
     if(chart){
-      chart.data.datasets[0].data.push(data1);
-      chart.data.datasets[1].data.push(data2);
+      console.log("UPDate chart is reactive on chart js.....");
+      chart.data.datasets[0].data.push(data.Phase1);
+      chart.data.datasets[1].data.push(data.Phase2);
       chart.data.labels.push(chart.data.labels.length + 1); 
       chart.update();
     }
@@ -66,7 +72,7 @@
 </script>
 
 <div style="display: flex; max-width: 50%;"> 
-  <canvas bind:this={chartCanvas} height="300"></canvas>
+  <canvas bind:this={chartCanvas} height=240></canvas>
 </div>
 
 <style>
