+++ Frontend/src/routes/Main/Dashboard/+page.svelte
@@ -1,10 +1,9 @@
 <script>
   import { onMount } from "svelte";
-  import { fetchData } from './stores';
   import Chart from "$lib/Components/Chart.svelte";
   
 
-  let data = [];
+  let data = {};
   let interval; 
 
   onMount(async () => {
@@ -19,27 +18,27 @@
 
  
 
-//  async function fetchData() {
+ async function fetchData() {
 
-//       try {
-//         const response = await fetch("http://localhost:8000/overview", {
-//       method: "POST", 
-//       headers: {
-//         'Content-Type': 'application/json' 
-//       }
+      try {
+        const response = await fetch("http://localhost:8000/overview", {
+      method: "POST", 
+      headers: {
+        'Content-Type': 'application/json' 
+      }
 
-//     });
-//         console.log("request being sent...");
-//         console.log(response);
-//         // const response = fetch("http://localhost:8000");
-//         const data = await response.json();
-//         console.log(data);
-//         //Voltage 1,2,3 as well as price
-//         return data; 
-//       } catch (error) {
-//         console.log("There was an error fetching the JSON for the overview..", error);
-//       }
-//   };
+    });
+        console.log("request being sent...");
+        console.log(response);
+        // const response = fetch("http://localhost:8000");
+        const fdata = await response.json();
+        console.log(data);
+        //Voltage 1,2,3 as well as price
+        data = fdata; 
+      } catch (error) {
+        console.log("There was an error fetching the JSON for the overview..", error);
+      }
+  };
 </script>
 
 <main class="container mx-auto">
@@ -74,7 +73,7 @@
     <p>Loading...</p>
   {/if}
 
-  <Chart class = "w-3/5 h-3/5"  />
+  <Chart class = "w-3/5 h-3/5" {data} />
